requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
coverage==5.2.1
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.2.1
django-filter==2.1.0
django-nine==0.2.3
django-nose==1.4.6
django-rest-auth==0.9.5
django-restframework==0.0.1
django-test-plus==1.4.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
importlib-metadata==1.7.0
iniconfig==1.0.1
ipaddress==1.0.23
more-itertools==8.4.0
nodeenv==1.5.0
nose==1.3.7
numpy==1.19.1
packaging==20.4
pandas==1.1.1
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
python-decouple==3.3
python-monkey-business==1.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
whitenoise==5.2.0
zipp==3.1.0